third_party/WebKit/LayoutTests/external/wpt/fetch/api/headers/headers-record-expected.txt

This is a testharness.js-based test.
PASS Passing nothing to Headers constructor 
FAIL Passing undefined to Headers constructor Failed to construct 'Headers': No matching constructor signature.
PASS Passing null to Headers constructor 
FAIL Basic operation with one property assert_equals: expected 4 but got 5
FAIL Basic operation with one property and a proto assert_equals: expected 4 but got 7
FAIL Correct operation ordering with two properties assert_equals: expected 6 but got 8
FAIL Correct operation ordering with two properties one of which has an invalid name assert_equals: expected 5 but got 8
FAIL Correct operation ordering with two properties one of which has an invalid value assert_equals: expected 4 but got 6
FAIL Correct operation ordering with non-enumerable properties assert_equals: expected 6 but got 7
FAIL Correct operation ordering with undefined descriptors assert_equals: expected 6 but got 7
FAIL Correct operation ordering with repeated keys assert_equals: expected 9 but got 10
Harness: the test ran to completion.
